PiOS Servo: remove old soft prescalar
Previous we got smart about setting the divisor
for very slow PWM rates. This had the same complexities
of inferring it for fast rates. Now we have the code to
set it explicitly, it is easy for someone to create a
new mode that uses 100khz as the time base, or whatever
is appropriate for their task.