Actuator: remove references to I2C ESCs
While refactoring the actuator settings, drop the
old I2C ESC support. No one has reported using it
in forever and it has probably bitrotted. This
drops the ChannelAddr field too, which was not
being used anywhere.

The ChannelAddr may need to return when we use CAN
ESCs but for now let's remove unused cruft.